Reorganize particles
