+++ SimulatorEngine/ParticleUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System.Numerics;
+using SimulatorEngine.Particles;
 
 namespace SimulatorEngine;
