+++ SimulatorEngine/Particles/Particle.cs
@@ -1,0 +1,32 @@
+﻿namespace SimulatorEngine.Particles;
+
+public enum ParticleKind
+{
+    None,
+    Sand,
+    Water,
+    Iron,
+    Oxygen,
+    Salt,
+    SaltyWater,
+}
+
+public enum ParticleBody
+{
+    Solid,
+    Powder,
+    Liquid,
+    Gas,
+}
+
+public abstract class Particle
+{
+    public int Temperature { get; set; }
+    public ParticleBody Body { get; set; }
+
+    public abstract ParticleKind GetKind();
+
+    public abstract uint GetColor();
+
+    public abstract float GetDensity();
+}
