+++ SimulatorEngine/ParticlesManager.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Numerics;
+using SimulatorEngine.Particles;
 
 namespace SimulatorEngine;
