+++ SimulatorUI/MainPage.xaml
@@ -3,7 +3,7 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:skia="clr-namespace:SkiaSharp.Views.Maui.Controls;assembly=SkiaSharp.Views.Maui.Controls"
-    xmlns:local="clr-namespace:SimulatorEngine;assembly=SimulatorEngine"
+    xmlns:local="clr-namespace:SimulatorEngine.Particles;assembly=SimulatorEngine"
     x:Class="SimulatorUI.MainPage"
     BackgroundColor="Beige">
 
@@ -50,11 +50,23 @@
                 WidthRequest="100"
                 BackgroundColor="#F6D7B0"/>
             <Button 
+                Text="Salt" 
+                Clicked="SetParticleKind"
+                CommandParameter="{x:Static local:ParticleKind.Salt}"
+                WidthRequest="100"
+                BackgroundColor="#FCF9F3"/>
+            <Button 
                 Text="Water" 
                 Clicked="SetParticleKind"
                 CommandParameter="{x:Static local:ParticleKind.Water}"
                 WidthRequest="100"
                 BackgroundColor="#1CA3EC"/>
+            <Button 
+                Text="Salt Water" 
+                Clicked="SetParticleKind"
+                CommandParameter="{x:Static local:ParticleKind.SaltyWater}"
+                WidthRequest="100"
+                BackgroundColor="#90AEBD"/>
             <Button 
                 Text="Iron" 
                 Clicked="SetParticleKind"
