+++ SimulatorUI/MainPage.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using SimulatorEngine;
+using SimulatorEngine.Particles;
 using SkiaSharp;
 using SkiaSharp.Views.Maui;
